Fix for windows shutdown otf listener crash.
CloseHandle on thread apparently doesn't actually terminate the thread.
Shutdown in the main thread causes the otf state to be freed out from under
the otf listener and can result in crashes. Here we signal otf shutdown
to the otf thread via the thread id/handle. It's possible that a more
principled approach is required (where we wait for proper thread shutdown).
Windows docs suggest we not explicitly kill the thread since there may be
dangling resources.  Testing and casual code inspection suggests that this
solution may be adequate.