Added "select" statement
This adds a new statement, "select". Before this addition it was
not possible to, say, chuck multiple events to `now` and resume on
the first event signalled, or to implement a timeout on waiting on
an event. Examples for these are in the also-added tests. Inspiration
for this feature comes of course from the Go language.

I believe I have also cleaned up some bits around event reference
counting, but I appreciate this is a bit of a twisty area and I could
well have missed something.