Revive the WINDOWS ASIO driver and enable multi-driver build.
- its desireable to ship a windows build with DS (Direct Sound) and ASIO
  support. A new commandline option, --driver:DS|ASIO|... allows the
  user to choose their preferred method.

We anticipate that the same multi-driver approach will work on linux
(and be a boon). (TBD).

Part of this change was to cleanup the use of WINDOWS_DS.  As I
understand it, this define represents DIRECTSOUND and should not be used
as an alias for WINDOWS builds.  It appears that PLATFORM_WIN32 is the
better name for that in the chuck context and this accounts for most of
the changes in the core.